% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{This function creates the folder structure for calibration requests}
\usage{
create_folder_structure(working_directory, aoi_name)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}
}
\value{
A folder called 'aoi_name' will be created with three folder in there: 'aoi', 'clustered', and 'results'.
}
\description{
This function creates the folder structure for calibration requests
}
\examples{
create_folder_structure(working_directory = 'D:/calibration/data/projects',
                        aoi_name = 'The_Netherlands')
}
