% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_threshold_raster.R
\name{create_threshold_raster}
\alias{create_threshold_raster}
\title{This creates a map of the AOI showing the distance if feature space with our calibration points}
\usage{
create_threshold_raster(
  working_directory,
  aoi_name,
  distance_threshold,
  current_calibration_points,
  country_shapefile,
  plot_resolution,
  plot_width,
  plot_heigth
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}

\item{distance_threshold}{The distance at which points can be considered a cluster (0.15 = default).}

\item{current_calibration_points}{The link (path + filename) to the file of our current calibration points with their corresponding covariable data.}

\item{country_shapefile}{Name of the admin_1 level of the country shapefile (e.g., 'NLD_adm1.shp').}

\item{plot_resolution}{Resolution for plotting of distance preview. 300 is default.}

\item{plot_width}{Plot width in pixels for plotting of distance preview. Set in range between 2000 - 6000.}

\item{plot_heigth}{Plot heigth in pixels  for plotting of distance preview. Set in range between 2000 - 6000.}
}
\value{
This creates a map of the AOI showing the distance in feature space with our calibration points. A raster and a preview of it will be returned. NOTE: this map can only be generated for countries countries that already have calibration points.
}
\description{
This creates a map of the AOI showing the distance if feature space with our calibration points
}
\examples{
create_threshold_raster(working_directory = 'D:/calibration/projects',
                        aoi_name = 'The_Netherlands',
                        country_code = 'NLA',
                        distance_threshold = 0.15,
                        current_calibration_points = 'D:/calibration/data/calibration_points/gps_covars.csv',
                        country_shapefile = 'D:/calibration/data/admin_regions/NLD_adm1.shp',
                        plot_resolution = 300,
                        plot_width = 4000,
                        plot_heigth = 3000)
}
