% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cropland_for_aoi.R,
%   R/get_cropland_for_manual_aoi.R
\name{get_cropland_for_aoi}
\alias{get_cropland_for_aoi}
\title{This function defines the to be calibrated area for administrative region}
\usage{
get_cropland_for_aoi(
  working_directory,
  aoi_name,
  vito_landcover_raster,
  base_raster_template,
  custom_aoi
)

get_cropland_for_aoi(
  working_directory,
  aoi_name,
  vito_landcover_raster,
  base_raster_template,
  custom_aoi
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}

\item{vito_landcover_raster}{The 2019 VITO landcover raster ('C:/data/vito_landcover_raster.tif')}

\item{base_raster_template}{Empty raster to reproject all data to ('C:/data/base_raster_template.tif')}

\item{custom_aoi}{A shapefile of a manually drawn AOI}

\item{country_shapefile}{Name of the admin_1 level of the country shapefile (e.g., 'NLD_adm1.shp')}

\item{target_regions}{A vector with names of target admin regions (e.g., c('Gelderland', 'Utrecht'))}

\item{complete_country}{Process the complete county (T/F). If T: target_regions will be ignored, if F, target_regions will be processed only.}

\item{cropland_only}{should all pixels in AOI be considered for calibration (set to FALSE), or only cropland pixels (set to TRUE)}
}
\value{
A folder called 'aoi_name' will be created with three folder in there: 'aoi', 'clustered', and 'results'.

A folder called 'aoi_name' will be created with three folder in there: 'aoi', 'clustered', and 'results'.
}
\description{
This function defines the to be calibrated area for administrative region

This function defines the to be calibrated area for a custom drawn AOI
}
\examples{
get_cropland_for_aoi(working_directory = 'C:/Users/peter/Documents',
                     aoi_name = 'The_Netherlands',
                     vito_landcover_raster = 'C:/data/vito_landcover_raster.tif',
                     base_raster_template = 'C:/data/base_raster_template.tif',
                     country_shapefile = 'NLD_adm1.shp',
                     target_regions = c('Gelderland', 'Utrecht'),
                     complete_country = TRUE)
get_cropland_for_aoi(working_directory = 'C:/Users/peter/Documents',
                     aoi_name = 'The_Netherlands',
                     vito_landcover_raster = 'C:/data/vito_landcover_raster.tif',
                     base_raster_template = 'C:/data/base_raster_template.tif',
                     custom_aoi = 'C:/data/custom_aoi.shp',
                     cropland_only = TRUE)
}
