% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cropland_for_custom_aoi.R
\name{get_cropland_for_custom_aoi}
\alias{get_cropland_for_custom_aoi}
\title{This function defines the to be calibrated area for a manually drawn AOI}
\usage{
get_cropland_for_custom_aoi(
  working_directory,
  aoi_name,
  vito_landcover_raster,
  base_raster_template,
  custom_aoi,
  cropland_only
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}

\item{vito_landcover_raster}{The 2019 VITO landcover raster ('D:/calibration/data/rasters/data/PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif')}

\item{base_raster_template}{Empty raster to reproject all data to ('D:/calibration/data/rasters/base_raster.tif')}

\item{custom_aoi}{A shapefile of a manually drawn AOI}

\item{cropland_only}{should all pixels in AOI be considered for calibration (set to FALSE), or only cropland pixels (set to TRUE)}
}
\value{
The to be calibrated area in for a manually drawn AOI will be defined and reprojected to a standard raster.
}
\description{
This function defines the to be calibrated area for a manually drawn AOI
}
\examples{
get_cropland_for_custom_aoi(working_directory = 'D:/calibration/data/projects',
                            aoi_name = 'The_Netherlands',
                            vito_landcover_raster = 'D:/calibration/data/rasters/data/PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif',
                            base_raster_template = 'D:/calibration/data/rasters/base_raster.tif',
                            custom_aoi = 'D:/calibration/data/shapefiles/custom_shapefile.shp',
                            cropland_only = TRUE)
}
