% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cropland_for_large_country.R
\name{get_cropland_for_large_country}
\alias{get_cropland_for_large_country}
\title{This function defines the to-be-calibrated-area for a large country}
\usage{
get_cropland_for_large_country(
  working_directory,
  aoi_name,
  vito_landcover_raster,
  base_raster_template,
  country_shapefile,
  cropland_only
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}

\item{vito_landcover_raster}{The 2019 VITO landcover raster ('D:/calibration/data/rasters/data/PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif').}

\item{base_raster_template}{Empty raster to reproject all data to ('D:/calibration/data/rasters/base_raster.tif').}

\item{country_shapefile}{Name of the admin_1 level of the country shapefile (e.g., 'NLD_adm1.shp').}

\item{cropland_only}{should all pixels in AOI be considered for calibration (set to FALSE), or only cropland pixels (set to TRUE).}
}
\value{
The to be calibrated area in for a large country will be defined and reprojected to a standard raster.
}
\description{
This function defines the to-be-calibrated-area for a large country
}
\examples{
get_cropland_for_large_country(working_directory = 'D:/calibration/projects',
                               aoi_name = 'The_Netherlands',
                               vito_landcover_raster = 'D:/calibration/data/rasters/data/PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif',
                               base_raster_template = 'D:/calibration/data/rasters/base_raster_template.tif',
                               country_shapefile = 'D:/calibration/data/admin_regions/NLD_adm1.shp',
                               cropland_only = TRUE)
}
