% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_of_calibration_points_kmeans.R
\name{get_location_of_calibration_points_kmeans}
\alias{get_location_of_calibration_points_kmeans}
\title{This function uses kmeans clustering to calculate the the location for a number of calibration points in an AOI}
\usage{
get_location_of_calibration_points_kmeans(
  working_directory,
  aoi_name,
  n_samples,
  iter.max,
  nstart,
  algorithm
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output is be stored. The previous steps of the calibration workflow have to be executed before running this.}

\item{aoi_name}{The name of the AOI as defined in the beginning of the workflow.}

\item{n_samples}{The number of calibration samples you want to collect in the AOI.}

\item{iter.max}{setting for KMeans clustering: maximum number of iterations.}

\item{nstart}{setting for KMeans clustering: number of starts.}

\item{algorithm}{setting for KMeans clustering: KMeans algorithm: choose one('Lloyd', 'Hartigan-Wong', 'Forgy', 'MacQueen').}
}
\value{
This function returns: 1) a .kml-file with the sampling locations named 1:n_samples, and 2) a raster with alternative sampling locations where each raster cell is numbered 1:n_samples.
}
\description{
This function uses kmeans clustering to calculate the the location for a number of calibration points in an AOI
}
\examples{
get_location_of_calibration_points_kmeans(working_directory = 'D:/calibration/projects',
                                          aoi_name = 'The_Netherlands',
                                          n_samples = 60,
                                          iter.max = 50,
                                          nstart = 10,
                                          algorithm = 'Lloyd')
}
