% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_covariable_data.R
\name{cluster_covariable_data}
\alias{cluster_covariable_data}
\title{This function clusters the covariable data in the aoi.}
\usage{
cluster_covariable_data(
  working_directory,
  aoi_name,
  distance_threshold,
  current_calibration_points
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}

\item{distance_threshold}{The distance at which points can be considered a cluster (0.15 = default).}

\item{current_calibration_points}{The link to a GPS file of our current calibration points.}
}
\value{
A folder called 'aoi_name' will be created with three folder in there: 'aoi', 'clustered', and 'results'.
}
\description{
This function clusters the covariable data in the aoi.
}
\examples{
cluster_covariable_data(working_directory = 'C:/Users/peter/Documents',
                        aoi_name = 'The_Netherlands',
                        distance_threshold = 0.15,
                        current_calibration_points = 'C:/Users/peter/Documents/calibration_gps.csv')
}
