% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calibration_curve.R
\name{create_calibration_curve}
\alias{create_calibration_curve}
\title{This function creates a calibration curve based on the clustered covariable data}
\usage{
create_calibration_curve(
  working_directory,
  aoi_name,
  country_shapefile,
  distance_threshold,
  within_threshold_distance,
  significant_cluster_size,
  legend_range,
  legend_location
)
}
\arguments{
\item{working_directory}{The directory where the folder calibration output will be stored. This directory should exist.}

\item{aoi_name}{The folder where the folder calibration output will be stored.}

\item{country_shapefile}{Name of the admin_1 level of the country shapefile (e.g., 'NLD_adm1.shp').}

\item{distance_threshold}{The distance at which points can be considered a cluster (0.15 = default).}

\item{within_threshold_distance}{The distance at which points can be considered inside a cluster (set to 1 for countries with calibration points and set to 'number of covariables' (which is currently 11) for countries without calibration points).}

\item{significant_cluster_size}{Ignore clusters smaller than this number.}

\item{legend_range}{Range of the legend (0 to legend_range). Guesstimate this based on size of country. Fine-tune with trial and error.}

\item{legend_location}{Where should the legend be plotted? 'topright', 'topleft', 'bottomright', 'bottomleft'. Fine-tune with trial and error.}
}
\value{
A plot with the calibration curve stored in the 'results' folder.
}
\description{
This function creates a calibration curve based on the clustered covariable data
}
\examples{
create_calibration_curve(working_directory = 'D:/calibration/data/projects',
                        aoi_name = 'The_Netherlands',
                        distance_threshold = 0.15,
                        country_shapefile = 'D:/calibration/data/admin_regions/NLD_adm1.shp',
                        distance_threshold = 0.15,
                        within_threshold_distance = 1,
                        significant_cluster_size = 3,
                        legend_range = 100,
                        legend_location = 'bottomleft')
}
